requirements.txt
openai
fastapi
uvicorn
jinja2
markdown
python-multipart
python-dotenv